update rehash script, add hash data for packages
